% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBreederInfo.R
\name{breederInfo}
\alias{breederInfo}
\title{R6 Class Representing a Breeder}
\description{
breederInfo object store specific information of one breeder.
}
\examples{

## ------------------------------------------------
## Method `breederInfo$new`
## ------------------------------------------------

### create simulation information
mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)

### create specie information
mySpec <- specie$new(nChr = 3,
                     lChr = c(100, 150, 200),
                     specName = "Example 1",
                     ploidy = 2,
                     mutRate = 10^-8,
                     recombRate = 10^-7,
                     recombRateOneVal = FALSE,
                     chrNames = c("C1", "C2", "C3"),
                     nLoci = 100,
                     effPopSize = 100,
                     simInfo = mySimInfo,
                     verbose = TRUE)
print(mySpec)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{breederName}}{[character] name of this breeder}

\item{\code{simInfo}}{[simInfo class] Simulation information
(see:\link[myBreedSimulatR]{simInfo})}

\item{\code{specie}}{[specie class] Specie of the SNPs
(see:\link[myBreedSimulatR]{specie})}

\item{\code{lociInfoFB}}{[list] information about the individuals haplotypes' SNPs available for breeder}

\item{\code{traitInfoFB}}{[list] Specific information of traits available for breeder}

\item{\code{popNameBase}}{[character] base of population's name.}

\item{\code{populationsFB}}{[list] A list of populationFb objects available for breeders}

\item{\code{crossInfoList}}{[list] A list of crossInfo objects}

\item{\code{generation}}{[list] current generation No. in the breeder}

\item{\code{calculateGRMBase}}{[logical] calculate genomic relationship matrix (GRM) for each population or not}

\item{\code{methodsGRMBase}}{[character] default methods to calculate GRM}

\item{\code{calcEpistasisBase}}{[logical] when additive / dominance GRM has already been calulated,
whether or not calculate epistatic GRM}

\item{\code{estimatedGVByGRMInfo}}{[list] A list of information on estimated GVs using GRM}

\item{\code{estimatedMrkEffInfo}}{[list] A list of information on estimated marker effects}

\item{\code{estimatedGVByMLRInfo}}{[list] A list of information on estimated GVs using MLR (multiple linear regression)}

\item{\code{verbose}}{[boolean] display information}

\item{\code{populationFB}}{[populationFB class] R6 class representing population for breeder}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{breederInfo$new()}}
\item \href{#method-getNewPopulation}{\code{breederInfo$getNewPopulation()}}
\item \href{#method-phenotyper}{\code{breederInfo$phenotyper()}}
\item \href{#method-estimateGVByGRM}{\code{breederInfo$estimateGVByGRM()}}
\item \href{#method-estimateMrkEff}{\code{breederInfo$estimateMrkEff()}}
\item \href{#method-estimateGVByMLR}{\code{breederInfo$estimateGVByMLR()}}
\item \href{#method-overGeneration}{\code{breederInfo$overGeneration()}}
\item \href{#method-removeLatestPop}{\code{breederInfo$removeLatestPop()}}
\item \href{#method-parentInd}{\code{breederInfo$parentInd()}}
\item \href{#method-lociEffects}{\code{breederInfo$lociEffects()}}
\item \href{#method-print}{\code{breederInfo$print()}}
\item \href{#method-clone}{\code{breederInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new specie object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{breederInfo$new(
  breederName = "Undefined",
  bsInfo,
  mrkNames = NULL,
  initGenotyping = TRUE,
  initGenotypedIndNames = NULL,
  mafThres = 0.05,
  heteroThres = 1,
  calculateGRMBase = TRUE,
  methodsGRMBase = "addNOIA",
  calcEpistasisBase = FALSE,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{breederName}}{[character] name of this breeder}

\item{\code{bsInfo}}{[bsInfo class] breeding scheme info (whichever generation is OK,
but it will use only 1st population)
  (see:\link[myBreedSimulatR]{bsInfo})}

\item{\code{mrkNames}}{[character] marker names}

\item{\code{initGenotyping}}{[logical] obtain marker genotype for initial population or not}

\item{\code{initGenotypedIndNames}}{[character] individual names that you want to genotype in initial population}

\item{\code{mafThres}}{[numeric] threshold for removing markers with maf < mafThres}

\item{\code{heteroThres}}{[numeric] threshold for removing markers with heteroRate >= heteroThres}

\item{\code{calculateGRMBase}}{[logical] calculate genomic relationship matrix (GRM) for each population or not}

\item{\code{methodsGRMBase}}{[character] default methods to calculate GRM}

\item{\code{calcEpistasisBase}}{[logical] when additive / dominance GRM has already been calulated,
whether or not calculate epistatic GRM}

\item{\code{verbose}}{[logical] Display info (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `breederInfo` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{### create simulation information
mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)

### create specie information
mySpec <- specie$new(nChr = 3,
                     lChr = c(100, 150, 200),
                     specName = "Example 1",
                     ploidy = 2,
                     mutRate = 10^-8,
                     recombRate = 10^-7,
                     recombRateOneVal = FALSE,
                     chrNames = c("C1", "C2", "C3"),
                     nLoci = 100,
                     effPopSize = 100,
                     simInfo = mySimInfo,
                     verbose = TRUE)
print(mySpec)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNewPopulation"></a>}}
\if{latex}{\out{\hypertarget{method-getNewPopulation}{}}}
\subsection{Method \code{getNewPopulation()}}{
get information on new population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{breederInfo$getNewPopulation(
  bsInfo,
  generationNew = NULL,
  genotyping = TRUE,
  genotypedIndNames = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bsInfo}}{[bsInfo class]  a bsInfo class object}

\item{\code{generationNew}}{[numeric] a generation of new population}

\item{\code{genotyping}}{[logical] Whether or not you want to genotype}

\item{\code{genotypedIndNames}}{[character] individual names that you want to genotype}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-phenotyper"></a>}}
\if{latex}{\out{\hypertarget{method-phenotyper}{}}}
\subsection{Method \code{phenotyper()}}{
get phenotypic values on current population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{breederInfo$phenotyper(
  bsInfo,
  generationOfInterest = NULL,
  nRep = 1,
  multiTraitsAsEnvs = FALSE,
  phenotypedIndNames = NULL,
  estimateGV = TRUE,
  estimatedGVMethod = "lme4"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bsInfo}}{[bsInfo class]  a bsInfo class object}

\item{\code{generationOfInterest}}{a generation where you want to obtain phenotypic values from population}

\item{\code{nRep}}{[numeric] Replication of the field trial (common to all traits)}

\item{\code{multiTraitsAsEnvs}}{[logical] Treat multiple traits as multiple environments or not}

\item{\code{phenotypedIndNames}}{[character] individual names that you want to phenotype}

\item{\code{estimateGV}}{[logical] estimate genotypic values by replication or not}

\item{\code{estimatedGVMethod}}{[character] We offer 'lme4' and 'mean'. 'lme4' is recommended.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGVByGRM"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGVByGRM}{}}}
\subsection{Method \code{estimateGVByGRM()}}{
estimate genotypic values based on GBLUP
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{breederInfo$estimateGVByGRM(
  trainingPop = NULL,
  testingPop = NULL,
  testingIndNames = NULL,
  methodsGRMFP = "addNOIA",
  bayesian = FALSE,
  multiTrait = FALSE,
  nIter = 10000,
  burnIn = 2000,
  thin = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trainingPop}}{[character / numeric] training population names or No.s (not generations!!)}

\item{\code{testingPop}}{[character / numeric] testing population names or No.s (not generations!!)}

\item{\code{testingIndNames}}{[character] names of testing individuals}

\item{\code{methodsGRMFP}}{[character] methods for calculating GRM for prediction}

\item{\code{bayesian}}{[logical] use bayesian model (BGLR) or not (RAINBOWR) for solving mixed-effects model}

\item{\code{multiTrait}}{[logical] use multiple-trait model for estimation of genotypic values}

\item{\code{nIter}}{[numeric] the number of iterations}

\item{\code{burnIn}}{[numeric] the number of burn-in}

\item{\code{thin}}{[numeric] the number of thinning}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateMrkEff"></a>}}
\if{latex}{\out{\hypertarget{method-estimateMrkEff}{}}}
\subsection{Method \code{estimateMrkEff()}}{
estimate marker effects based on multiple linear regression (machine learning)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{breederInfo$estimateMrkEff(
  trainingPop = NULL,
  methodMLR = "Ridge",
  multiTrait = FALSE,
  alpha = 0.5,
  nIter = 10000,
  burnIn = 2000,
  thin = 5,
  bayesian = FALSE,
  alphaMarker = 0.5,
  sizeMrkMin = 2,
  sizeMrkMax = 12
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trainingPop}}{[character / numeric] training population names or No.s (not generations!!)}

\item{\code{methodMLR}}{[character] methods for estimating marker effects.
The following methods are offered:

"Ridge", "LASSO", "ElasticNet", "RR-BLUP", "GBLUP", "BayesA" (uni-trait),
"BayesB" (uni-trait), "BayesC" (uni-trait), "BRR", "BL" (uni-trait), "SpikeSlab" (multi-trait)}

\item{\code{multiTrait}}{[logical] use multiple-trait model for estimation of marker effects or not}

\item{\code{alpha}}{[numeric] the elastic net mixing parameter, with \eqn{0 \leq \alpha \leq 1}.
The penalty is defined as

\eqn{\frac {1 - \alpha} { 2 } || \beta || _ 2 ^ 2 + \alpha || \beta || _ 1}

`alpha = 1` is the lasso penalty, and `alpha = 0` the ridge penalty.}

\item{\code{nIter}}{[numeric] the number of iterations}

\item{\code{burnIn}}{[numeric] the number of burn-in}

\item{\code{thin}}{[numeric] the number of thinning}

\item{\code{bayesian}}{[logical] use bayesian model (BGLR) or not (RAINBOWR) for solving mixed-effects model
(only when `methodMLR = 'GBLUP'`, this argument is valid.)}

\item{\code{alphaMarker}}{[numeric] for plot: transparency for markers, see \link[plotly]{plot_ly}}

\item{\code{sizeMrkMin}}{[numeric] for plot: size for marker with minimum estimated effect}

\item{\code{sizeMrkMax}}{[numeric] for plot: size for marker with maximum estimated effect}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGVByMLR"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGVByMLR}{}}}
\subsection{Method \code{estimateGVByMLR()}}{
estimate genotypic values based on GBLUP
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{breederInfo$estimateGVByMLR(
  trainingPop = NULL,
  testingPop = NULL,
  testingIndNames = NULL,
  methodMLR = "Ridge",
  multiTrait = FALSE,
  alpha = 0.5,
  nIter = 10000,
  burnIn = 2000,
  thin = 5,
  bayesian = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trainingPop}}{[character / numeric] training population names or No.s (not generations!!)}

\item{\code{testingPop}}{[character / numeric] testing population names or No.s (not generations!!)}

\item{\code{testingIndNames}}{[character] names of testing individuals}

\item{\code{methodMLR}}{[character] methods for estimating marker effects.
The following methods are offered:

"Ridge", "LASSO", "ElasticNet", "RR-BLUP", "GBLUP", "BayesA" (uni-trait),
"BayesB" (uni-trait), "BayesC" (uni-trait), "BRR", "BL" (uni-trait), "SpikeSlab" (multi-trait)}

\item{\code{multiTrait}}{[logical] use multiple-trait model for estimation of marker effects or not}

\item{\code{alpha}}{[numeric] the elastic net mixing parameter, with \eqn{0 \leq \alpha \leq 1}.
The penalty is defined as

\eqn{\frac {1 - \alpha} { 2 } || \beta || _ 2 ^ 2 + \alpha || \beta || _ 1}

`alpha = 1` is the lasso penalty, and `alpha = 0` the ridge penalty.}

\item{\code{nIter}}{[numeric] the number of iterations}

\item{\code{burnIn}}{[numeric] the number of burn-in}

\item{\code{thin}}{[numeric] the number of thinning}

\item{\code{bayesian}}{[logical] use bayesian model (BGLR) or not (RAINBOWR) for solving mixed-effects model
(only when `methodMLR = 'GBLUP'`, this argument is valid.)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-overGeneration"></a>}}
\if{latex}{\out{\hypertarget{method-overGeneration}{}}}
\subsection{Method \code{overGeneration()}}{
Display informations about the object


assemble populations over generation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{breederInfo$overGeneration(targetPop = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{targetPop}}{[character / numeric] population names or No.s you want to assemble.
If NULL, assemble all the populations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removeLatestPop"></a>}}
\if{latex}{\out{\hypertarget{method-removeLatestPop}{}}}
\subsection{Method \code{removeLatestPop()}}{
remove latest population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{breederInfo$removeLatestPop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parentInd"></a>}}
\if{latex}{\out{\hypertarget{method-parentInd}{}}}
\subsection{Method \code{parentInd()}}{
search parents of an individual of interest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{breederInfo$parentInd(indName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indName}}{[character] individual name of interest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lociEffects"></a>}}
\if{latex}{\out{\hypertarget{method-lociEffects}{}}}
\subsection{Method \code{lociEffects()}}{
search parents of an individual of interest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{breederInfo$lociEffects(bsInfo, trainingPop = NULL, methodMLR = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bsInfo}}{[bsInfo class] breeding scheme info (whichever generation is OK,
but it will use only 1st population)
  (see:\link[myBreedSimulatR]{bsInfo})#'}

\item{\code{trainingPop}}{[character / numeric] training population names or No.s (not generations!!)}

\item{\code{methodMLR}}{[character] methods for estimating marker effects.
The following methods are offered:

"Ridge", "LASSO", "ElasticNet", "RR-BLUP", "GBLUP", "BayesA" (uni-trait),
"BayesB" (uni-trait), "BayesC" (uni-trait), "BRR", "BL" (uni-trait), "SpikeSlab" (multi-trait)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Display information about the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{breederInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{breederInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
