% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHaplotype.R
\name{haplotype}
\alias{haplotype}
\title{R6 Class representing an haplotype}
\description{
haplotype object store specific information of individuals haplotype
}
\examples{

## ------------------------------------------------
## Method `haplotype$new`
## ------------------------------------------------

### create simulation information
mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)

### create specie information
mySpec <- specie$new(nChr = 3,
                     lChr = c(100, 150, 200),
                     specName = "Example 1",
                     ploidy = 2,
                     mutRate = 10^-8,
                     recombRate = 10^-6,
                     recombRateOneVal = FALSE,
                     chrNames = c("C1", "C2", "C3"),
                     nLoci = c(3, 4, 5),
                     effPopSize = 3,
                     simInfo = mySimInfo,
                     verbose = TRUE)

### create lociInfo object
myLoci <- lociInfo$new(genoMap = NULL, specie = mySpec)

### simulate haplotype
rawHaplo <- matrix(sample(c(0, 1), (3 + 4 + 5) * 3, replace = TRUE),
                   nrow = 2)
colnames(rawHaplo) <- paste0("Locus_", 1:(3 + 4 + 5))

myHaplo <- myBreedSimulatR::haplotype$new(lociInfo = myLoci,
                                          haplo = rawHaplo)
myHaplo$values
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lociInfo}}{[lociInfo class] information about the haplotype of the
SNPs (see:\link[myBreedSimulatR]{lociInfo})}

\item{\code{values}}{[list] list haplotype values}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{allelDose}}{[numeric] vector of haplotypes encoded in allele dose}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{haplotype$new()}}
\item \href{#method-clone}{\code{haplotype$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Haplotype object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{haplotype$new(lociInfo, haplo)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lociInfo}}{[lociInfo class] information about the haplotype's SNPs
(see:\link[myBreedSimulatR]{lociInfo})}

\item{\code{haplo}}{[matrix] named matrix of the genotype for all markers}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `lociInfo` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{### create simulation information
mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)

### create specie information
mySpec <- specie$new(nChr = 3,
                     lChr = c(100, 150, 200),
                     specName = "Example 1",
                     ploidy = 2,
                     mutRate = 10^-8,
                     recombRate = 10^-6,
                     recombRateOneVal = FALSE,
                     chrNames = c("C1", "C2", "C3"),
                     nLoci = c(3, 4, 5),
                     effPopSize = 3,
                     simInfo = mySimInfo,
                     verbose = TRUE)

### create lociInfo object
myLoci <- lociInfo$new(genoMap = NULL, specie = mySpec)

### simulate haplotype
rawHaplo <- matrix(sample(c(0, 1), (3 + 4 + 5) * 3, replace = TRUE),
                   nrow = 2)
colnames(rawHaplo) <- paste0("Locus_", 1:(3 + 4 + 5))

myHaplo <- myBreedSimulatR::haplotype$new(lociInfo = myLoci,
                                          haplo = rawHaplo)
myHaplo$values
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{haplotype$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
