% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStoSOO.R
\name{layer}
\alias{layer}
\title{R6 Class Representing a layer in tree for StoSOO}
\description{
layer object store specific information on each layer in tree for StoSOO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nodesList}}{[list] List of nodes (`node` class object) in this layer}

\item{\code{depth}}{[numeric] Depth of this layer}

\item{\code{optimizeType}}{[character] Either 'deterministic' for optimizing a deterministic function or 'stochastic' for a stochastic one}

\item{\code{nChildrenPerExpansion}}{[numeric] Number of children per expansion}

\item{\code{nNodes}}{[numeric] Number nodes in this layer}

\item{\code{positions}}{[numeric] Positions of the nodes in this layer}

\item{\code{ucbValues}}{[numeric] Upper confidence bounds of the nodes in this layer}

\item{\code{rewardMeans}}{[numeric] Empirical averages of `rewards` of the nodes in this layer}

\item{\code{isLeaves}}{[numeric] Are the nodes in this layer leaves of the tree or not}

\item{\code{withCheck}}{[logical] Check arguments for `node`, `layer`, and `tree` class or not}

\item{\code{maxUcbValues}}{[numeric] Maximum UCB value (or `rewardMean` for 'deterministic' function) of the leaf node in the layer}

\item{\code{positionMaxUcbValues}}{[numeric] Position of the leaf node which shows maximum UCB value (or `rewardMean` for 'deterministic' function)}

\item{\code{maxUcbValuesSoFar}}{[numeric] Maximum UCB value for the layers shallower than this layer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{layer$new()}}
\item \href{#method-addNodes}{\code{layer$addNodes()}}
\item \href{#method-sortedByPosition}{\code{layer$sortedByPosition()}}
\item \href{#method-evaluateMaximum}{\code{layer$evaluateMaximum()}}
\item \href{#method-performOneAction}{\code{layer$performOneAction()}}
\item \href{#method-print}{\code{layer$print()}}
\item \href{#method-clone}{\code{layer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new layer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{layer$new(
  nodesList,
  depth = NULL,
  optimizeType,
  nChildrenPerExpansion,
  withCheck = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nodesList}}{[list] List of nodes (`node` class object) in this layer}

\item{\code{depth}}{[numeric] Depth of this layer}

\item{\code{optimizeType}}{[character] Either 'deterministic' for optimizing a deterministic function or 'stochastic' for a stochastic one}

\item{\code{nChildrenPerExpansion}}{[numeric] Number of children per expansion}

\item{\code{withCheck}}{[logical] Check arguments for `node`, `layer`, and `tree` class or not}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addNodes"></a>}}
\if{latex}{\out{\hypertarget{method-addNodes}{}}}
\subsection{Method \code{addNodes()}}{
Add new nodes to the layer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{layer$addNodes(newNodesList = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newNodesList}}{[list] List of new nodes (`node` class object) to be added to this layer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sortedByPosition"></a>}}
\if{latex}{\out{\hypertarget{method-sortedByPosition}{}}}
\subsection{Method \code{sortedByPosition()}}{
Sort nodes in this layer by position
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{layer$sortedByPosition()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evaluateMaximum"></a>}}
\if{latex}{\out{\hypertarget{method-evaluateMaximum}{}}}
\subsection{Method \code{evaluateMaximum()}}{
Evaluate maximum value of the leaf node in this layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{layer$evaluateMaximum()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-performOneAction"></a>}}
\if{latex}{\out{\hypertarget{method-performOneAction}{}}}
\subsection{Method \code{performOneAction()}}{
Update layer information once
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{layer$performOneAction(maxUcbValuesSoFar)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxUcbValuesSoFar}}{[numeric] Maximum UCB value for the layers shallower than this layer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Display information about the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{layer$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{layer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
