% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLoci.R
\name{lociInfo}
\alias{lociInfo}
\title{R6 Class representing a set of SNP markers}
\description{
lociInfo object store specific information of a set of SNP
markers including QTLs
}
\details{
This class is useful for setting the Haplotype class. (see:
\code{\link[myBreedSimulatR]{haplotype}})
}
\examples{

## ------------------------------------------------
## Method `lociInfo$new`
## ------------------------------------------------

mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)

### create specie information
mySpec <- specie$new(nChr = 3,
                     lChr = c(100, 150, 200),
                     specName = "Example 1",
                     ploidy = 2,
                     mutRate = 10^-8,
                     recombRate = 10^-7,
                     recombRateOneVal = FALSE,
                     chrNames = c("C1", "C2", "C3"),
                     nLoci = 100,
                     effPopSize = 100,
                     simInfo = mySimInfo,
                     verbose = TRUE)

### create SNP information
myLoci <- lociInfo$new(genoMap = NULL,
                     specie = mySpec,
                     seedSimGeno = NA,
                     lociNames = NULL,
                     founderNames = NULL)
print(myLoci)


## ------------------------------------------------
## Method `lociInfo$nLoci`
## ------------------------------------------------

myLoci$nLoci()
myLoci$nLoci(c("C2","C3"))

## ------------------------------------------------
## Method `lociInfo$getInfo`
## ------------------------------------------------

myLoci$getInfo("Loci_1")
myLoci$getInfo(c("Loci_1", "Loci_3"))

## ------------------------------------------------
## Method `lociInfo$plot`
## ------------------------------------------------

myLoci$plot(alpha = 0.1)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{genoMap}}{[data.frame] Coordinate of all SNPs.

3 columns:
\itemize{
 \item{\code{lociNames}:} {SNP&QTL's IDs}
 \item{\code{chr}:} {Chromosome holding the SNP}
 \item{\code{pos}:} {SNP position on the chromosome}
 }}

\item{\code{specie}}{[specie class] Specie of the SNPs
(see:\link[myBreedSimulatR]{specie})}

\item{\code{posGeneticDist}}{[logical] Position is based on genetic distance or not (physical distance).
If you simulate genome data, this argument will automatically be TRUE.
If this argument is FALSE, `specie$recombRateOneVal` will be TRUE.}

\item{\code{seedSimGeno}}{[numeric] Seed for simulating marker genotype}

\item{\code{founderHaplo}}{[matrix] Haplotype of founders (\eqn{n \times m} matrix) by simulation (for real data, founderHaplo = NULL)}

\item{\code{genoMapList}}{[list] Named list of dataframes with the coordinate
of all SNPs. for each chromosomes}

\item{\code{ids}}{[list] Named list of the SNP ids for all chromosomes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{lociInfo$new()}}
\item \href{#method-nLoci}{\code{lociInfo$nLoci()}}
\item \href{#method-getInfo}{\code{lociInfo$getInfo()}}
\item \href{#method-print}{\code{lociInfo$print()}}
\item \href{#method-plot}{\code{lociInfo$plot()}}
\item \href{#method-clone}{\code{lociInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new lociInfo object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lociInfo$new(
  genoMap = NULL,
  specie,
  posGeneticDist = TRUE,
  seedSimGeno = NA,
  lociNames = NULL,
  founderNames = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genoMap}}{[data.frame] Coordinate of all SNPs. (needed for real data)

  3 columns: \itemize{ \item{\code{chr}:} {Chromosome holding the SNP}
  \item{\code{pos}:} {SNP position on the chromosome}
  \item{\code{lociNames}:} {SNP's IDs} }}

\item{\code{specie}}{[specie class] Specie of the SNPs
(see:\link[myBreedSimulatR]{specie})}

\item{\code{posGeneticDist}}{[logical] Position is based on genetic distance or not (physical distance).
If you simulate genome data, this argument will automatically be TRUE.
If this argument is FALSE, `specie$recombRateOneVal` will be TRUE.}

\item{\code{seedSimGeno}}{[numeric] Random seed for simulation of marker genotype (optional)}

\item{\code{lociNames}}{[character] Names of the markers used for simulated marker genotype (optional)}

\item{\code{founderNames}}{[character] Names of the founders. It should be equal to ploidy * effective population size (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `lociInfo` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)

### create specie information
mySpec <- specie$new(nChr = 3,
                     lChr = c(100, 150, 200),
                     specName = "Example 1",
                     ploidy = 2,
                     mutRate = 10^-8,
                     recombRate = 10^-7,
                     recombRateOneVal = FALSE,
                     chrNames = c("C1", "C2", "C3"),
                     nLoci = 100,
                     effPopSize = 100,
                     simInfo = mySimInfo,
                     verbose = TRUE)

### create SNP information
myLoci <- lociInfo$new(genoMap = NULL,
                     specie = mySpec,
                     seedSimGeno = NA,
                     lociNames = NULL,
                     founderNames = NULL)
print(myLoci)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nLoci"></a>}}
\if{latex}{\out{\hypertarget{method-nLoci}{}}}
\subsection{Method \code{nLoci()}}{
Get the number of SNPs per chromosomes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lociInfo$nLoci(chr = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chr}}{[str or numeric] chromosome id}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myLoci$nLoci()
myLoci$nLoci(c("C2","C3"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInfo"></a>}}
\if{latex}{\out{\hypertarget{method-getInfo}{}}}
\subsection{Method \code{getInfo()}}{
Get information about specific SNPs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lociInfo$getInfo(lociNames)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lociNames}}{[str] loci ids}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myLoci$getInfo("Loci_1")
myLoci$getInfo(c("Loci_1", "Loci_3"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Display summary information about the object: specie, number
of SNP, SNP coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lociInfo$print(fh = TRUE, fl = TRUE, rown = 6, coln = 6, rowst = 1, colst = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fh}}{From head. If this argument is TRUE, first part (row) of data will be shown (like head() function).
If FALSE, last part (row) of your data will be shown (like tail() function).}

\item{\code{fl}}{From left. If this argument is TRUE, first part (column) of data will be shown (like head() function).
If FALSE, last part (column) of your data will be shown (like tail() function).}

\item{\code{rown}}{The number of rows shown in console.}

\item{\code{coln}}{The number of columns shown in console.}

\item{\code{rowst}}{The start point for the direction of row.}

\item{\code{colst}}{The start point for the direction of column.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
plot chromosome map using the \pkg{plotly} package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lociInfo$plot(alpha = 0.01)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{transparency see \link[plotly]{plot_ly}}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myLoci$plot(alpha = 0.1)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lociInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
