% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCross.R
\name{makeSingleCross}
\alias{makeSingleCross}
\title{Cross two individuals together}
\usage{
makeSingleCross(ind1, ind2, names, n = 1, verbose = TRUE)
}
\arguments{
\item{ind1}{parent 1}

\item{ind2}{parent 2}

\item{names}{names of the descendants}

\item{n}{number of descendants}

\item{verbose}{print informations}
}
\value{
list of new individuals
}
\description{
Cross two individuals together
}
\examples{
### create simulation information
mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)

### create specie information
mySpec <- specie$new(nChr = 3,
                     lChr = c(100, 150, 200),
                     specName = "Example 1",
                     ploidy = 2,
                     mutRate = 10^-8,
                     recombRate = 10^-6,
                     chrNames = c("C1", "C2", "C3"),
                     nLoci = c(3, 4, 5),
                     recombRateOneVal = FALSE,
                     effPopSize = 3,
                     simInfo = mySimInfo,
                     verbose = TRUE)

### create lociInfo object
myLoci <- lociInfo$new(genoMap = NULL, specie = mySpec)

### simulate haplotype
rawHaplo1 <- matrix(sample(c(0, 1), (3 + 4 + 5) * 2, replace = TRUE),
                    nrow = 2)
colnames(rawHaplo1) <- paste0("Locus_", 1:(3 + 4 + 5))

myHaplo1 <- myBreedSimulatR::haplotype$new(lociInfo = myLoci,
                                           haplo = rawHaplo1)

rawHaplo2 <- matrix(sample(c(0, 1), (3 + 4 + 5) * 2, replace = TRUE),
                    nrow = 2)
colnames(rawHaplo2) <- paste0("Locus_", 1:(3 + 4 + 5))

myHaplo2 <- myBreedSimulatR::haplotype$new(lociInfo = myLoci,
                                           haplo = rawHaplo2)

### create individuals:
myInd1 <-  individual$new(name = "Ind 1",
                          specie = mySpec,
                          parent1 = "OkaaSan1",
                          parent2 = "OtouSan1",
                          haplo = myHaplo1,
                          verbose = TRUE)

myInd2 <-  individual$new(name = "Ind 2",
                          specie = mySpec,
                          parent1 = "OkaaSan2",
                          parent2 = "OtouSan2",
                          haplo = myHaplo2,
                          verbose = TRUE)
offspring <- makeSingleCross(myInd1, myInd2, names = "off 1")
offspring
}
