% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySelection.R
\name{selectWBV}
\alias{selectWBV}
\title{Selection according to weighted breeding values}
\usage{
selectWBV(pop, n, lociEffects)
}
\arguments{
\item{pop}{(population object) Population of individuals to select
(see:\link[breedSimulatR]{population})}

\item{n}{(integer) number of individuals to select}

\item{lociEffects}{(numeric vector) effect of the genetic markers}
}
\value{
character vector of the selected individuals' names
}
\description{
Selection according to weighted breeding values
}
\examples{
mySpec <- specie$new(specName = "Statisticae exempli",
                     nChr = 10,
                     lChr = 1e6,
                     ploidy = 2,
                     recombRate = 3/1e6)
myLoci <- lociInfo$new(genoMap = exampleData$genoMap, specie = mySpec)
initPop <- createPop(geno = exampleData$genotypes,
                     lociInfo = myLoci,
                     popName = "Initial population")
selectWBV(pop = initPop,
         lociEffects = exampleData$lociEffects,
         n = 10)
}
\references{
Jannink, Jean-Luc. “Dynamics of Long-Term Genomic Selection.”
  Genetics Selection Evolution 42, no. 1 (December 2010).
  https://doi.org/10.1186/1297-9686-42-35.
}
