% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySimInfo.R
\name{simInfo}
\alias{simInfo}
\title{R6 Class representing a simulation information}
\description{
simInfo object store specific information of one simulation study.
}
\details{
Details: simInfo object store specific information of one simulation study.
}
\examples{

## ------------------------------------------------
## Method `simInfo$new`
## ------------------------------------------------

mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)
print(mySimInfo)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simName}}{[character] Name of the simulation study}

\item{\code{simGeno}}{[logical] Simulate marker genotype or not}

\item{\code{simPheno}}{[logical] Simulate phenotypic values or not}

\item{\code{nSimGeno}}{[numeric] Number of simulations for generating marker genome}

\item{\code{nSimPheno}}{[numeric] Number of simulations for pseudo phenotypic values}

\item{\code{nCoreMax}}{[numeric] Number of cores you can use for this simulation study}

\item{\code{nCorePerGeno}}{[numeric] Number of cores used per simulation of marker genotype (optional)}

\item{\code{nCorePerPheno}}{[numeric] Number of cores used per simulation of phenotypic values (optional)}

\item{\code{saveDataFileName}}{[character] The name of the file (including path) that saves the data (if NULL, the data won't be saved)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{simInfo$new()}}
\item \href{#method-print}{\code{simInfo$print()}}
\item \href{#method-clone}{\code{simInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new simInfo object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simInfo$new(
  simName = "Simulation Example",
  simGeno = TRUE,
  simPheno = TRUE,
  nSimGeno = NA,
  nSimPheno = NA,
  nCoreMax = NA,
  nCorePerGeno = NA,
  nCorePerPheno = NA,
  saveDataFileName = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simName}}{[character] Name of the simulation study}

\item{\code{simGeno}}{[logical] Simulate marker genotype or not}

\item{\code{simPheno}}{simPheno [logical] Simulate phenotypic values or not}

\item{\code{nSimGeno}}{[numeric] Number of simulations for generating marker genome (optional)}

\item{\code{nSimPheno}}{[numeric] Number of simulations for pseudo phenotypic values (optional)}

\item{\code{nCoreMax}}{[numeric] Number of cores you can use for this simulation study (optional)}

\item{\code{nCorePerGeno}}{[numeric] Number of cores used per simulation of marker genotype (optional)}

\item{\code{nCorePerPheno}}{[numeric] Number of cores used per simulation of phenotypic values (optional)}

\item{\code{saveDataFileName}}{[character] The name of the file (including path) that saves the data (if NULL, the data won't be saved)
(optional)}

\item{\code{verbose}}{[logical] Display info (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `simInfo` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)
print(mySimInfo)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Display information about the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
