% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTrait.R
\name{traitInfo}
\alias{traitInfo}
\title{R6 Class representing information of traits (especially simulated ones)}
\description{
traitInfo object store specific information of traits (especially simulated ones).
}
\examples{

## ------------------------------------------------
## Method `traitInfo$new`
## ------------------------------------------------

mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)

### create specie information
mySpec <- specie$new(nChr = 3,
                     lChr = c(100, 150, 200),
                     specName = "Example 1",
                     ploidy = 2,
                     mutRate = 10^-8,
                     recombRate = 10^-7,
                     recombRateOneVal = FALSE,
                     chrNames = c("C1", "C2", "C3"),
                     nLoci = 100,
                     effPopSize = 100,
                     simInfo = mySimInfo,
                     verbose = TRUE)

### create SNP information
myLoci <- lociInfo$new(genoMap = NULL,
                     specie = mySpec,
                     seedSimGeno = NA,
                     lociNames = NULL,
                     founderNames = NULL)


### create traitInfo object
myTrait <- traitInfo$new(lociInfo = myLoci,
                         nMarkers = 80,
                         nTraits = 3,
                         nQTLs = c(4, 8, 3),
                         actionTypeEpiSimple = TRUE,
                         qtlOverlap = TRUE,
                         nOverlap = c(2, 3, 0),
                         effCor = 0.1,
                         propDomi = 0.2,
                         interactionMean = c(4, 1, 2))
print(myTrait)


## ------------------------------------------------
## Method `traitInfo$genoMapDetail`
## ------------------------------------------------

myLoci$genoMapDetail()
myLoci$genoMapDetail("Loci_6")

## ------------------------------------------------
## Method `traitInfo$plot`
## ------------------------------------------------

myLoci$plot(alphaMarker = 0.1, alphaQTL = 0.9,
            sizeMarker = 5, sizeQTLmin = 6, sizeQTLmax = 12)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lociInfo}}{[lociInfo class] lociInfo of the specie}

\item{\code{nMarkers}}{[numeric] Number of markers for each chromosome}

\item{\code{nTraits}}{[numeric] Number of traits assumed in this breeding scheme}

\item{\code{traitNames}}{[character] Names of traits}

\item{\code{nQTLs}}{[matrix] Number of QTLs of each chromosome for each trait}

\item{\code{qtlVarList}}{[list] A list of variances of QTLs for additive/dominance, pleiotropy, and epistasis

\itemize{
 \item{\code{$qtlPle}:} {A variance of QTLs for pleiotropy}
 \item{\code{$qtlEach}:} {A variance of QTLs for each trait (except epistasis)}
 \item{\code{$qtlEpi}:} {A variance of QTLs for each trait (epistasis)}
 }}

\item{\code{qtlOverlap}}{[logical] If TRUE, pleiotropy across traits will be assumed}

\item{\code{nOverlap}}{[numeric] Number of QTLs common across traits for each chromosome}

\item{\code{effCor}}{[numeric] Correlation of pleiotropic QTL effects between traits}

\item{\code{propDomi}}{[numeric] Proportion of dominant effects for each trait}

\item{\code{interactionMean}}{[numeric] The expected number that each QTL interacts with others for each trait}

\item{\code{actionTypeEpiSimple}}{[logical] If TRUE, `actionTypeEpi` will be same as
`actionType` for each QTL.}

\item{\code{seedSelectMarker}}{[numeric] Random seed for selecting marker information}

\item{\code{seedSelectQTLEach}}{[numeric] Random seed for selecting each QTL and its effect (it should have the same length as nTraits)}

\item{\code{seedSelectQTLPle}}{[numeric] Random seed for selecting pleiotropic QTLs and their effects}

\item{\code{seedSelectQTLDom}}{[numeric] Random seed for selecting QTL positions with dominant effects (it should have the same length as nTraits)}

\item{\code{seedSelectQTLEpi}}{[numeric] Random seed for selecting epistasis and its effect (it should have the same length as nTraits)}

\item{\code{mrkPos}}{[numeric] Marker positions}

\item{\code{mrkPosEachChr}}{[list] Marker positions for each chromosome}

\item{\code{qtlPos}}{[list] QTL positions for each trait}

\item{\code{qtlEachPos}}{[list] Positions of QTLs specific to one trait for each trait}

\item{\code{qtlPlePos}}{[list] Pleiotropic QTL positions for each trait}

\item{\code{qtlEpiPos}}{[list] Epistatic QTL positions for each trait}

\item{\code{qtlPosEachChrList}}{[list] QTL positions for each trait for each chromosome}

\item{\code{qtlEachPosEachChrList}}{[list] Positions of QTLs specific to one trait for each trait for each chromosome}

\item{\code{qtlPlePosEachChrList}}{[list] Pleiotropic QTL positions for each trait for each chromosome}

\item{\code{qtlNamesList}}{[list] QTL names for each trait}

\item{\code{qtlEachNamesList}}{[list] Names of QTLs common to one trait for each trait}

\item{\code{qtlPleNamesList}}{[list] Pleiotropic QTL names for each trait}

\item{\code{qtlEpiNamesList}}{[list] Epistatic QTL names for each trait}

\item{\code{qtlAllNamesList}}{[list] All QTL names for each trait (including epistatic effects)}

\item{\code{actionType}}{[list] A list of vectors representing additive by 0 and dominance by 1 for each trait}

\item{\code{actionTypeEpi}}{[list] A list of matrix representing additive by 0 and dominance by 1 for epitasis for each trait}

\item{\code{qtlEff}}{[list] QTL effects for each trait}

\item{\code{qtlPleEff}}{[list] Pleiotropic QTL effects for each trait}

\item{\code{qtlEachEff}}{[list] Effects of QTLs specific to one trait for each trait}

\item{\code{qtlEpiEff}}{[list] Epistatic QTL effects for each trait}

\item{\code{qtlAllEff}}{[list] All QTL effects for each trait (including epistasis)}

\item{\code{nQTLsEach}}{[numeric] Number of QTLs specific to each trait}

\item{\code{nEpi}}{[numeric] Number of epistatic effects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{traitInfo$new()}}
\item \href{#method-genoMapDetail}{\code{traitInfo$genoMapDetail()}}
\item \href{#method-print}{\code{traitInfo$print()}}
\item \href{#method-plot}{\code{traitInfo$plot()}}
\item \href{#method-clone}{\code{traitInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new traitInfo object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{traitInfo$new(
  lociInfo,
  nMarkers,
  nTraits = NULL,
  traitNames = NULL,
  nQTLs = NULL,
  qtlVarList = NULL,
  qtlOverlap = TRUE,
  nOverlap = NULL,
  effCor = NULL,
  propDomi = NULL,
  interactionMean = NULL,
  actionTypeEpiSimple = TRUE,
  qtlPleNamesList = NULL,
  qtlEachNamesList = NULL,
  seedSelectMarker = NA,
  seedSelectQTLEach = NA,
  seedSelectQTLPle = NA,
  seedSelectQTLDom = NA,
  seedSelectQTLEpi = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lociInfo}}{[lociInfo class] lociInfo of the specie}

\item{\code{nMarkers}}{[numeric] Number of markers for each chromosome}

\item{\code{nTraits}}{[numeric] Number of traits assumed in this breeding scheme}

\item{\code{traitNames}}{[character] Names of traits}

\item{\code{nQTLs}}{[matrix] Number of QTLs of each chromosome for each trait}

\item{\code{qtlVarList}}{[list] A list of variances of QTLs for additive/dominance, pleiotropy, and epistasis

\itemize{
 \item{\code{$qtlPle}:} {A variance of QTLs for pleiotropy}
 \item{\code{$qtlEach}:} {A variance of QTLs for each trait (except epistasis)}
 \item{\code{$qtlEpi}:} {A variance of QTLs for each trait (epistasis)}
 }}

\item{\code{qtlOverlap}}{[logical] If TRUE, pleiotropy across traits will be assumed}

\item{\code{nOverlap}}{[numeric] Number of QTLs common across traits for each chromosome}

\item{\code{effCor}}{[numeric] Correlation of pleiotropic QTL effects between traits}

\item{\code{propDomi}}{[numeric] Proportion of dominant effects for each trait}

\item{\code{interactionMean}}{[numeric] The expected number that each QTL interacts with others for each trait}

\item{\code{actionTypeEpiSimple}}{[logical] If TRUE, `actionTypeEpi` will be same as
`actionType` for each QTL.}

\item{\code{qtlPleNamesList}}{[list] Pleiotropic QTL names for each trait}

\item{\code{qtlEachNamesList}}{[list] Names of QTLs common to one trait for each trait}

\item{\code{seedSelectMarker}}{[numeric] Random seed for selecting marker information}

\item{\code{seedSelectQTLEach}}{[numeric] Random seed for selecting each QTL and its effect (it should have the same length as nTraits)}

\item{\code{seedSelectQTLPle}}{[numeric] Random seed for selecting pleiotropic QTLs and their effects}

\item{\code{seedSelectQTLDom}}{[numeric] Random seed for selecting QTL positions with dominant effects (it should have the same length as nTraits)}

\item{\code{seedSelectQTLEpi}}{[numeric] Random seed for selecting epistasis and its effect (it should have the same length as nTraits)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `traitInfo` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)

### create specie information
mySpec <- specie$new(nChr = 3,
                     lChr = c(100, 150, 200),
                     specName = "Example 1",
                     ploidy = 2,
                     mutRate = 10^-8,
                     recombRate = 10^-7,
                     recombRateOneVal = FALSE,
                     chrNames = c("C1", "C2", "C3"),
                     nLoci = 100,
                     effPopSize = 100,
                     simInfo = mySimInfo,
                     verbose = TRUE)

### create SNP information
myLoci <- lociInfo$new(genoMap = NULL,
                     specie = mySpec,
                     seedSimGeno = NA,
                     lociNames = NULL,
                     founderNames = NULL)


### create traitInfo object
myTrait <- traitInfo$new(lociInfo = myLoci,
                         nMarkers = 80,
                         nTraits = 3,
                         nQTLs = c(4, 8, 3),
                         actionTypeEpiSimple = TRUE,
                         qtlOverlap = TRUE,
                         nOverlap = c(2, 3, 0),
                         effCor = 0.1,
                         propDomi = 0.2,
                         interactionMean = c(4, 1, 2))
print(myTrait)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-genoMapDetail"></a>}}
\if{latex}{\out{\hypertarget{method-genoMapDetail}{}}}
\subsection{Method \code{genoMapDetail()}}{
Detail information of map for marker genome
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{traitInfo$genoMapDetail(lociNames = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lociNames}}{[character] loci ids}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myLoci$genoMapDetail()
myLoci$genoMapDetail("Loci_6")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Display summary information about the object: traitInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{traitInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
plot QTL map using the \pkg{plotly} package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{traitInfo$plot(
  alphaMarker = 0.05,
  alphaQTL = 0.8,
  sizeMarker = 5,
  sizeQTLmin = 6,
  sizeQTLmax = 12
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alphaMarker}}{[numeric] transparency for markers see \link[plotly]{plot_ly}}

\item{\code{alphaQTL}}{[numeric] transparency for QTLs see \link[plotly]{plot_ly}}

\item{\code{sizeMarker}}{[numeric] size for markers}

\item{\code{sizeQTLmin}}{[numeric] size for QTL with minimum effect}

\item{\code{sizeQTLmax}}{[numeric] size for QTL with maximum effect}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myLoci$plot(alphaMarker = 0.1, alphaQTL = 0.9,
            sizeMarker = 5, sizeQTLmin = 6, sizeQTLmax = 12)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{traitInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
