% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTrial.R
\name{trialInfo}
\alias{trialInfo}
\title{R6 Class representing information of field trial}
\description{
traitInfo object store specific information of field trial.
}
\examples{

## ------------------------------------------------
## Method `trialInfo$new`
## ------------------------------------------------

### create simulation information
mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)
### create specie information
mySpec <- specie$new(nChr = 3,
                     lChr = c(100, 150, 200),
                     specName = "Example 1",
                     ploidy = 2,
                     mutRate = 10^-8,
                     recombRate = 10^-6,
                     chrNames = c("C1", "C2", "C3"),
                     nLoci = 100,
                     recombRateOneVal = FALSE,
                     effPopSize = 100,
                     simInfo = mySimInfo,
                     verbose = TRUE)
### create lociInfo object
myLoci <- lociInfo$new(genoMap = NULL, specie = mySpec)
plot(myLoci, alpha = 0.1)
### create traitInfo object
myTrait <- traitInfo$new(lociInfo = myLoci,
                         nMarkers = 80,
                         nTraits = 3,
                         nQTLs = c(4, 8, 3),
                         actionTypeEpiSimple = TRUE,
                         qtlOverlap = TRUE,
                         nOverlap = c(2, 3, 0),
                         effCor = 0.1,
                         propDomi = 0.2,
                         interactionMean = c(4, 1, 2))
### create simulated population
simulatedPop <- createPop(geno = NULL,
                          haplo = NULL,
                          lociInfo = myLoci,
                          traitInfo = myTrait,
                          founderIsInitPop = TRUE,
                          popName = "First Population",
                          verbose = FALSE)
### create trialInfo object
myTrial <- trialInfo$new(population = simulatedPop,
                         herit = c(0.4, 0.7, 0.2),
                         nRep = 2,
                         multiTraitsAsEnvs = FALSE,
                         residCor = 0.3)
simulatedPop$inputTrialInfo(trialInfoNow = myTrial)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{population}}{[population class] population class object}

\item{\code{herit}}{[numeric] Heritability for each trait (plot-based/line-based)}

\item{\code{nRep}}{[numeric] Replication of the field trial (common to all traits)}

\item{\code{multiTraitsAsEnvs}}{[logical] Treat multiple traits as multiple environments or not}

\item{\code{envSpecificEffects}}{[numeric] Effects specific to each environments / treatments.
If `multiTraitsAsEnvs = FALSE`, envSpecificEffects will be 0 for all traits.}

\item{\code{residCor}}{[matrix] Residual correlation between traits}

\item{\code{seedResid}}{[numeric] Random seed for selecting residuals}

\item{\code{trueGeneticVar}}{[numeric] True genetic variance}

\item{\code{trueGeneticCov}}{[matrix] True genetic covariance between traits}

\item{\code{trueGeneticCor}}{[matrix] True genetic correlation between traits}

\item{\code{residVar}}{[numeric] Residual variance for each trait}

\item{\code{residCov}}{[matrix] Residual covariance between traits}

\item{\code{resid}}{[array] individual x traits x replication (3-dimensional array) of residuals}

\item{\code{trueResidVar}}{[numeric] True residual variance for each trait}

\item{\code{trueResidCov}}{[matrix] True residual covariance between traits}

\item{\code{trueResidCor}}{[matrix] True residual correlation between traits}

\item{\code{trueHeritLine}}{[numeric] True heritability of each trait (plot-based/line-based)}

\item{\code{trueHeritInd}}{[numeric] True heritability of each trait (individual-based)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{trialInfo$new()}}
\item \href{#method-print}{\code{trialInfo$print()}}
\item \href{#method-clone}{\code{trialInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new trialInfo object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trialInfo$new(
  population,
  herit = NULL,
  nRep = NULL,
  multiTraitsAsEnvs = FALSE,
  envSpecificEffects = NULL,
  residCor = NULL,
  seedResid = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{population}}{[population class] population class object}

\item{\code{herit}}{[numeric] Heritability for each trait (plot-based/line-based)}

\item{\code{nRep}}{[numeric] Replication of the field trial (common to all traits)}

\item{\code{multiTraitsAsEnvs}}{[logical] Treat multiple traits as multiple environments or not}

\item{\code{envSpecificEffects}}{[numeric] Effects specific to each environments / treatments.
If `multiTraitsAsEnvs = FALSE`, envSpecificEffects will be 0 for all traits.}

\item{\code{residCor}}{[matrix] Residual correlation between traits}

\item{\code{seedResid}}{[numeric] Random seed for selecting residuals}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `trialInfo` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{### create simulation information
mySimInfo <- simInfo$new(simName = "Simulation Example",
                         simGeno = TRUE,
                         simPheno = TRUE,
                         nSimGeno = 1,
                         nSimPheno = 3,
                         nCoreMax = 4,
                         nCorePerGeno = 1,
                         nCorePerPheno = 3,
                         saveDataFileName = NULL)
### create specie information
mySpec <- specie$new(nChr = 3,
                     lChr = c(100, 150, 200),
                     specName = "Example 1",
                     ploidy = 2,
                     mutRate = 10^-8,
                     recombRate = 10^-6,
                     chrNames = c("C1", "C2", "C3"),
                     nLoci = 100,
                     recombRateOneVal = FALSE,
                     effPopSize = 100,
                     simInfo = mySimInfo,
                     verbose = TRUE)
### create lociInfo object
myLoci <- lociInfo$new(genoMap = NULL, specie = mySpec)
plot(myLoci, alpha = 0.1)
### create traitInfo object
myTrait <- traitInfo$new(lociInfo = myLoci,
                         nMarkers = 80,
                         nTraits = 3,
                         nQTLs = c(4, 8, 3),
                         actionTypeEpiSimple = TRUE,
                         qtlOverlap = TRUE,
                         nOverlap = c(2, 3, 0),
                         effCor = 0.1,
                         propDomi = 0.2,
                         interactionMean = c(4, 1, 2))
### create simulated population
simulatedPop <- createPop(geno = NULL,
                          haplo = NULL,
                          lociInfo = myLoci,
                          traitInfo = myTrait,
                          founderIsInitPop = TRUE,
                          popName = "First Population",
                          verbose = FALSE)
### create trialInfo object
myTrial <- trialInfo$new(population = simulatedPop,
                         herit = c(0.4, 0.7, 0.2),
                         nRep = 2,
                         multiTraitsAsEnvs = FALSE,
                         residCor = 0.3)
simulatedPop$inputTrialInfo(trialInfoNow = myTrial)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Display summary information about the object: traitInfo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trialInfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trialInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
